requirements.txt
flask==2.3.3
pdf2image==1.16.3
Pillow==10.0.1
openpyxl==3.1.2
requests==2.31.0
python-dotenv==1.0.0
azure-ai-vision==0.13.0
azure-ai-formrecognizer==3.3.0
azure-cognitiveservices-vision-computervision==0.9.0
azure-ai-textanalytics==5.3.0 